+++ packages/storage/index.ts
@@ -1,13 +1,10 @@
 export * from './CookieStorageAdapter';
 export * from './ExpiableStorageAdapter';
 export * from './MemorySimpleStorage';
-export * from './RefreshableJWTProvider';
 export * from './SimpleStorageAdapter';
 export * from './StorageTokenProvider';
 export * from './cookie';
-export * from './createJWTProvider.factory';
 export * from './memoryStorage';
 export * from './createStorage.factory';
 export * from './storage.type';
-export * from './storage.util';
 export * from './createTokenProvider.factory';
